resource content_version
This is separate to CrossMark, unless you particularly want it to display in the record tab, in which case you should include it in your CrossMark deposit. It’s not needed to make CrossMark work.
It would be too difficult to add the urls for each of the different versions of an article in an automated fashion, so deleting this from the sample as it is not required