Crossmark_domain_exclusive
Changed from true to false
If your content is likely to legitimately appear in many places on different websites then you can choose to be “non domain-exclusive” and not register any CrossMark domains. When a reader clicks on the CrossMark logo on any of these other websites the CrossMark metadata will be retrieved from CrossRef and displayed to the reader. If there are updates available, the reader will be directed to the publisher’s website (via the CrossRef DOI) to view the updates.